+++ application_sdk/server/fastapi/models.py
@@ -5,7 +5,7 @@ from typing import Any, Dict, List, Optional, Type
 
 from pydantic import BaseModel, Field, RootModel
 
-from application_sdk.outputs.eventstore import Event, EventFilter
+from application_sdk.events.models import Event, EventFilter
 from application_sdk.workflows import WorkflowInterface
 
 
@@ -34,7 +34,7 @@ class FetchMetadataRequest(RootModel[Dict[str, Any]]):
 
 class FetchMetadataResponse(BaseModel):
     success: bool
-    data: List[Dict[str, str]]
+    data: Any
 
 
 class PreflightCheckRequest(BaseModel):